+++ jdaviz/configs/imviz/plugins/orientation/orientation.vue
@@ -3,7 +3,7 @@
     :config="config"
     plugin_key="Orientation"
     :api_hints_enabled.sync="api_hints_enabled"
-    :description="docs_description || 'Rotate the viewer orientation or choose to align images by pixels.'"
+    :description="docs_description"
     :link="docs_link || 'https://jdaviz.readthedocs.io/en/'+vdocs+'/'+config+'/plugins.html#imviz-orientation'"
     :popout_button="popout_button"
     :scroll_to.sync="scroll_to"